removes units from root collection, restructures civil shared project
