+++ Connectors/Autocad/Speckle.Connectors.AutocadShared/Operations/Send/AutocadRootObjectBuilder.cs
@@ -7,7 +7,6 @@ using Speckle.Connectors.Common.Caching;
 using Speckle.Connectors.Common.Conversion;
 using Speckle.Connectors.Common.Extensions;
 using Speckle.Connectors.Common.Operations;
-using Speckle.Converters.Autocad;
 using Speckle.Converters.Common;
 using Speckle.Sdk;
 using Speckle.Sdk.Logging;
@@ -21,7 +20,6 @@ public class AutocadRootObjectBuilder : IRootObjectBuilder<AutocadRootObject>
 {
   private readonly IRootToSpeckleConverter _converter;
   private readonly string[] _documentPathSeparator = ["\\"];
-  private readonly IConverterSettingsStore<AutocadConversionSettings> _converterSettings;
   private readonly ISendConversionCache _sendConversionCache;
   private readonly AutocadInstanceUnpacker _instanceUnpacker;
   private readonly AutocadMaterialUnpacker _materialUnpacker;
@@ -40,8 +38,7 @@ public class AutocadRootObjectBuilder : IRootObjectBuilder<AutocadRootObject>
     AutocadLayerUnpacker layerUnpacker,
     AutocadGroupUnpacker groupUnpacker,
     ILogger<AutocadRootObjectBuilder> logger,
-    ISdkActivityFactory activityFactory,
-    IConverterSettingsStore<AutocadConversionSettings> converterSettings
+    ISdkActivityFactory activityFactory
   )
   {
     _converter = converter;
@@ -53,7 +50,6 @@ public class AutocadRootObjectBuilder : IRootObjectBuilder<AutocadRootObject>
     _groupUnpacker = groupUnpacker;
     _logger = logger;
     _activityFactory = activityFactory;
-    _converterSettings = converterSettings;
   }
 
   public Task<RootObjectBuilderResult> Build(
@@ -89,7 +85,6 @@ public class AutocadRootObjectBuilder : IRootObjectBuilder<AutocadRootObject>
           .Reverse()
           .First()
       };
-    root["units"] = _converterSettings.Current.SpeckleUnits;
 
     // TODO: better handling for document and transactions!!
     Document doc = Application.DocumentManager.CurrentDocument;
