+++ Connectors/Autocad/Speckle.Connectors.AutocadShared/Plugin/AutocadCommand.cs
@@ -31,7 +31,7 @@ public class AutocadCommand
       return;
     }
 
-    PaletteSet = new PaletteSet("Speckle (Beta) for Autocad", s_id)
+    PaletteSet = new PaletteSet($"Speckle (Beta) for {AppUtils.App}", s_id)
     {
       Size = new Size(400, 500),
       DockEnabled = (DockSides)((int)DockSides.Left + (int)DockSides.Right)
@@ -52,7 +52,7 @@ public class AutocadCommand
 
     var panelWebView = Container.GetRequiredService<DUI3ControlWebView>();
 
-    PaletteSet.AddVisual("Speckle (Beta) for Autocad WebView", panelWebView);
+    PaletteSet.AddVisual($"Speckle (Beta) for {AppUtils.App} WebView", panelWebView);
 
     FocusPalette();
   }
